@@ -11,20 +11,32 @@
     "clear": "docusaurus clear",
     "serve": "docusaurus serve",
     "write-translations": "docusaurus write-translations",
-    "write-heading-ids": "docusaurus write-heading-ids"
+    "write-heading-ids": "docusaurus write-heading-ids",
+    "typecheck": "tsc"
   },
   "dependencies": {
-    "@docusaurus/core": "2.4.3",
-    "@docusaurus/preset-classic": "2.4.3",
-    "@mdx-js/react": "^1.6.22",
+    "@docusaurus/core": "0.0.0-5709",
+    "@docusaurus/preset-classic": "0.0.0-5709",
+    "@docusaurus/remark-plugin-npm2yarn": "0.0.0-5709",
+    "@docusaurus/theme-live-codeblock": "0.0.0-5709",
+    "@docusaurus/theme-mermaid": "0.0.0-5709",
+    "@mdx-js/react": "^2.3.0",
     "clsx": "^1.2.1",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "hast-util-is-element": "1.1.0",
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "rehype-katex": "5",
+    "remark-math": "3",
+    "swagger-ui-react": "^5.9.0"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "2.4.3",
-    "dotenv": "^16.3.1"
+    "@docusaurus/module-type-aliases": "0.0.0-5709",
+    "@docusaurus/tsconfig": "0.0.0-5709",
+    "@docusaurus/types": "0.0.0-5709",
+    "@types/swagger-ui-react": "^4.18.1",
+    "dotenv": "^16.3.1",
+    "typescript": "~5.2.2"
   },
   "browserslist": {
     "production": [
@@ -33,12 +45,12 @@
       "not op_mini all"
     ],
     "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
+      "last 3 chrome version",
+      "last 3 firefox version",
+      "last 5 safari version"
     ]
   },
   "engines": {
-    "node": ">=16.14"
+    "node": ">=18.0"
   }
 }
